build 1.0.0
- Finally null safe
- Dependencies are now up-to date
