+++ example/.metadata
@@ -4,7 +4,7 @@
 # This file should be version controlled and should not be manually edited.
 
 version:
-  revision: 8af6b2f038c1172e61d418869363a28dffec3cb4
+  revision: 7e9793dee1b85a243edd0e06cb1658e98b077561
   channel: stable
 
 project_type: app
