+++ example/android/settings.gradle
@@ -1,7 +1,3 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
 include ':app'
 
 def localPropertiesFile = new File(rootProject.projectDir, "local.properties")
